GEODE-11: Add a basic score based aggregator
* Aggregator merges results returned from shards and orders the result based on
  lucene hit score
* A configurable count of results is merged

Review link: https://reviews.apache.org/r/37720/

+++ gemfire-lucene/src/main/java/com/gemstone/gemfire/cache/lucene/LuceneResultStruct.java
@@ -1,6 +1,5 @@
 package com.gemstone.gemfire.cache.lucene;
 
-import java.util.LinkedList;
 
 /**
  * <p>
@@ -41,7 +40,7 @@ public interface LuceneResultStruct {
    * @return score
    * @throws IllegalArgumentException If this struct does not contain score
    */
-  public Double getScore();
+  public float getScore();
   
   /**
    * Get the types of values ordered list
